cli: move speak into a growable audio group (nativ audio speak)
TTS is one kind of audio-out; group it under `nativ audio` so other tasks
(separation, enhancement, sfx) can land as thin reflections of future server
endpoints. ttsModel config unchanged; agent reference + README updated.
