+++ cli/Sources/nativ/Commands/Agent.swift
@@ -67,8 +67,9 @@ struct Agent: AsyncParsableCommand {
       array-of-vectors. Reads stdin (one input per line) if no args.
     - `image [prompt...] [--out file] [--size WxH]` — generate an image to a file.
     - `transcribe <audio-file>` — speech-to-text, prints the transcript.
-    - `speak [text...] [--out file] [--voice v] [--speed s] [--format mp3]` —
-      text-to-speech; writes an audio file (reads stdin if no text).
+    - `audio speak [text...] [--out file] [--voice v] [--speed s] [--format mp3]` —
+      text-to-speech; writes an audio file (reads stdin if no text). The `audio`
+      group is where other audio-out tasks land as the server grows endpoints.
     - `config show | set | path` — read/write cli.json.
     - `agent [--json]` — this reference.
 
@@ -115,7 +116,7 @@ struct Agent: AsyncParsableCommand {
                 ["name": "embed", "desc": "text embeddings", "flags": ["--model", "--dims"]],
                 ["name": "image", "desc": "generate an image", "flags": ["--model", "--out", "--size"]],
                 ["name": "transcribe", "desc": "speech-to-text", "flags": ["--model"]],
-                ["name": "speak", "desc": "text-to-speech", "flags": ["--model", "--out", "--voice", "--speed", "--format"]],
+                ["name": "audio speak", "desc": "text-to-speech (audio group grows: separate/enhance/sfx)", "flags": ["--model", "--out", "--voice", "--speed", "--format"]],
                 ["name": "config", "desc": "show/set cli.json", "flags": []],
             ],
             "api": [
